refactor: 패키지 구조 변경, api 엔드 포인트 변경
