+++ algofi-compile/src/main/java/gooroommoon/algofi_compile/judge/exception/RequestException.java
@@ -1,8 +1,10 @@
-package gooroommoon.algofi_compile.exception;
+package gooroommoon.algofi_compile.judge.exception;
+
+import gooroommoon.algofi_compile.input.InputJudgeControllerAdvice;
 
 /**
  * 이 예외가 발생했다는 것은 사용자가 잘못된 요청을 보냈음을 의미하므로 400 Bad Request 반환
- * @see ApiControllerAdvice
+ * @see InputJudgeControllerAdvice
  */
 public class RequestException extends RuntimeException {
     public RequestException() {
