+++ algofi-compile/src/main/java/gooroommoon/algofi_compile/judge/exception/ServerException.java
@@ -1,8 +1,10 @@
-package gooroommoon.algofi_compile.exception;
+package gooroommoon.algofi_compile.judge.exception;
+
+import gooroommoon.algofi_compile.input.InputJudgeControllerAdvice;
 
 /**
  * 이 예외가 발생했다는 것은 파일 시스템이나 동시성 문제가 발생했을 수 있으므로 501 Server Error 반환
- * @see ApiControllerAdvice
+ * @see InputJudgeControllerAdvice
  */
 public class ServerException extends RuntimeException {
     public ServerException() {
