+++ algofi-compile/src/main/java/gooroommoon/algofi_compile/judge/service/language/PythonCodeExecutor.java
@@ -1,16 +1,16 @@
-package gooroommoon.algofi_compile.service.language;
+package gooroommoon.algofi_compile.judge.service.language;
 
-import gooroommoon.algofi_compile.exception.ServerException;
+import gooroommoon.algofi_compile.judge.exception.ServerException;
 
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
-public class JavaScriptCodeExecutor implements CodeExecutor {
+public class PythonCodeExecutor implements CodeExecutor {
     @Override
     public Path makeFileFromCode(String code) {
-        Path path = Paths.get("temp.js");
+        Path path = Paths.get("temp.py");
 
         try {
             Files.write(path, code.getBytes());
@@ -23,6 +23,6 @@ public class JavaScriptCodeExecutor implements CodeExecutor {
 
     @Override
     public ProcessBuilder executeProcessBuilder(Path path) {
-        return new ProcessBuilder("sh", "-c", "node " + path);
+        return new ProcessBuilder("sh", "-c", "python3 " + path);
     }
 }
